Add async tests for GZipStream
