+++ src/System.IO.Compression/tests/DeflateStreamTests.cs
@@ -863,222 +863,4 @@ namespace System.IO.Compression.Tests
             await base.WriteAsync(array, offset, count, cancellationToken);
         }
     }
-
-    public abstract class DeflateStreamAsyncTestsBase
-    {
-        public abstract Task<int> ReadAsync(DeflateStream unzip, byte[] buffer, int offset, int count);
-        public abstract Task WriteAsync(DeflateStream unzip, byte[] buffer, int offset, int count);
-        protected static string gzTestFile(string fileName) { return Path.Combine("GZTestData", fileName); }
-
-        [Fact]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework, "Full Framework Flush is a no-op.")]
-        public async Task OverlappingFlushAsync_DuringFlushAsync()
-        {
-            byte[] buffer = null;
-            string testFilePath = gzTestFile("GZTestDocument.pdf");
-            using (var origStream = await LocalMemoryStream.readAppFileAsync(testFilePath))
-            {
-                buffer = origStream.ToArray();
-            }
-
-            using (var writeStream = new ManualSyncMemoryStream(false))
-            using (var zip = new DeflateStream(writeStream, CompressionMode.Compress))
-            {
-                Task task = null;
-                try
-                {
-                    writeStream.manualResetEvent.Set();
-                    await WriteAsync(zip, buffer, 0, buffer.Length);
-                    writeStream.manualResetEvent.Reset();
-                    writeStream.WriteHit = false;
-                    task = zip.FlushAsync();
-                    Assert.True(writeStream.WriteHit);
-                    Assert.Throws<InvalidOperationException>(() => { zip.FlushAsync(); }); // "overlapping flushes"
-                }
-                finally
-                {
-                    // Unblock Async operations
-                    writeStream.manualResetEvent.Set();
-                    // The original WriteAsync should be able to complete
-                    Assert.True(task.Wait(100 * 500));
-                }
-            }
-        }
-
-        [Fact]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework, "Full Framework Flush is a no-op.")]
-        public async Task OverlappingFlushAsync_DuringWriteAsync()
-        {
-            byte[] buffer = null;
-            string testFilePath = gzTestFile("GZTestDocument.pdf");
-            using (var origStream = await LocalMemoryStream.readAppFileAsync(testFilePath))
-            {
-                buffer = origStream.ToArray();
-            }
-
-            using (var writeStream = new ManualSyncMemoryStream(false))
-            using (var zip = new DeflateStream(writeStream, CompressionMode.Compress))
-            {
-                Task task = null;
-                try
-                {
-                    task = WriteAsync(zip, buffer, 0, buffer.Length);
-                    Assert.True(writeStream.WriteHit);
-                    Assert.Throws<InvalidOperationException>(() => { zip.FlushAsync(); }); // "overlapping flushes"
-                }
-                finally
-                {
-                    // Unblock Async operations
-                    writeStream.manualResetEvent.Set();
-                    // The original WriteAsync should be able to complete
-                    Assert.True(task.Wait(100 * 500));
-                }
-            }
-        }
-
-        [Fact]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework, "Full Framework Flush is a no-op.")]
-        public async Task OverlappingWriteAsync()
-        {
-            byte[] buffer = null;
-            string testFilePath = gzTestFile("GZTestDocument.pdf");
-            using (var origStream = await LocalMemoryStream.readAppFileAsync(testFilePath))
-            {
-                buffer = origStream.ToArray();
-            }
-
-            using (var writeStream = new ManualSyncMemoryStream(false))
-            using (var zip = new DeflateStream(writeStream, CompressionMode.Compress))
-            {
-                Task task = null;
-                try
-                {
-                    task = WriteAsync(zip, buffer, 0, buffer.Length);    // write needs to be bigger than the internal write buffer
-                    Assert.True(writeStream.WriteHit);
-                    Assert.Throws<InvalidOperationException>(() => { zip.WriteAsync(buffer, 32, 32); }); // "overlapping write"
-                }
-                finally
-                {
-                    // Unblock Async operations
-                    writeStream.manualResetEvent.Set();
-                    // The original WriteAsync should be able to complete
-                    Assert.True(task.Wait(100 * 500));
-                }
-            }
-        }
-
-        [Fact]
-        [SkipOnTargetFramework(TargetFrameworkMonikers.NetFramework, "Full Framework Flush is a no-op.")]
-        public async Task OverlappingReadAsync()
-        {
-            byte[] buffer = new byte[32];
-            string testFilePath = gzTestFile("GZTestDocument.pdf.gz");
-            using (var readStream = await ManualSyncMemoryStream.GetStreamFromFileAsync(testFilePath, false, true))
-            using (var unzip = new DeflateStream(readStream, CompressionMode.Decompress, true))
-            {
-                Task<int> task = null;
-                try
-                {
-                    task = ReadAsync(unzip, buffer, 0, 32);
-                    Assert.True(readStream.ReadHit);
-                    Assert.Throws<InvalidOperationException>(() => { ReadAsync(unzip, buffer, 0, 32); }); // "overlapping read"
-                }
-                finally
-                {
-                    // Unblock Async operations
-                    readStream.manualResetEvent.Set();
-                    // The original ReadAsync should be able to complete
-                    Assert.True(task.Wait(100 * 500));
-                }
-            }
-        }
-
-        [Fact]
-        public async Task OverlappingFlushAsync_DuringReadAsync()
-        {
-            byte[] buffer = new byte[32];
-            string testFilePath = gzTestFile("GZTestDocument.pdf.gz");
-            using (var readStream = await ManualSyncMemoryStream.GetStreamFromFileAsync(testFilePath, false, true))
-            using (var unzip = new DeflateStream(readStream, CompressionMode.Decompress, true))
-            {
-                Task task = null;
-                try
-                {
-                    task = ReadAsync(unzip, buffer, 0, 32);
-                    Assert.True(readStream.ReadHit);
-                    Assert.Throws<InvalidOperationException>(() => { unzip.FlushAsync(); }); // "overlapping read"
-                }
-                finally
-                {
-                    // Unblock Async operations
-                    readStream.manualResetEvent.Set();
-                    // The original ReadAsync should be able to complete
-                    Assert.True(task.Wait(100 * 500));
-                }
-            }
-        }
-
-        [Fact]
-        public async Task DecompressWorks()
-        {
-            var compareStream = await LocalMemoryStream.readAppFileAsync(gzTestFile("GZTestDocument.txt"));
-            var gzStream = await LocalMemoryStream.readAppFileAsync(gzTestFile("GZTestDocument.txt.gz"));
-
-            await DecompressAsync(compareStream, gzStream);
-        }
-
-        [Fact]
-        public async Task DecompressWorksWithBinaryFile()
-        {
-            var compareStream = await LocalMemoryStream.readAppFileAsync(gzTestFile("GZTestDocument.doc"));
-            var gzStream = await LocalMemoryStream.readAppFileAsync(gzTestFile("GZTestDocument.doc.gz"));
-
-            await DecompressAsync(compareStream, gzStream);
-        }
-
-        // Making this async since regular read/write are tested below
-        private async Task DecompressAsync(MemoryStream compareStream, MemoryStream gzStream)
-        {
-            var strippedMs = StripHeaderAndFooter.Strip(gzStream);
-
-            var ms = new MemoryStream();
-            var zip = new DeflateStream(strippedMs, CompressionMode.Decompress);
-
-            var deflateStream = new MemoryStream();
-
-            int _bufferSize = 1024;
-            var bytes = new byte[_bufferSize];
-            bool finished = false;
-            int retCount;
-            while (!finished)
-            {
-                retCount = await ReadAsync(zip, bytes, 0, _bufferSize);
-
-                if (retCount != 0)
-                    await deflateStream.WriteAsync(bytes, 0, retCount);
-                else
-                    finished = true;
-            }
-
-            deflateStream.Position = 0;
-            compareStream.Position = 0;
-
-            byte[] compareArray = compareStream.ToArray();
-            byte[] writtenArray = deflateStream.ToArray();
-
-            Assert.Equal(compareArray.Length, writtenArray.Length);
-            for (int i = 0; i < compareArray.Length; i++)
-            {
-                Assert.Equal(compareArray[i], writtenArray[i]);
-            }
-        }
-    }
-
-    public sealed class TaskDeflateStreamTests : DeflateStreamAsyncTestsBase
-    {
-        public override Task<int> ReadAsync(DeflateStream unzip, byte[] buffer, int offset, int count) => unzip.ReadAsync(buffer, offset, count);
-
-        public override Task WriteAsync(DeflateStream unzip, byte[] buffer, int offset, int count) => unzip.WriteAsync(buffer, offset, count);
-    }
-
 }
